requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
python-dotenv==1.0.0
httpx==0.25.2
aiosqlite==0.19.0
